requirements.txt
appdirs==1.4.0
dj-database-url==0.4.2
Django==1.10.5
django-oidc-provider==0.4.4
ecdsa==0.13
future==0.16.0
gevent==1.2.1
greenlet==0.4.12
gunicorn==19.6.0
packaging==16.8
pycrypto==2.6.1
pycryptodomex==3.4.3
pyjwkest==1.3.0
pyparsing==2.1.10
python-jose==1.3.2
PyYAML==3.12
requests==2.13.0
six==1.10.0
